Match dbsh_symbol_update
Force both modelState flag clears to a materialized-constant mask via a
u64 cast (#74: li -5; and), keeping obj and modelState in separate
registers as the target does.

+++ src/main/dll/dll_0196_dbshsymbol.c
@@ -246,7 +246,7 @@ void dbsh_symbol_update(int obj)
         phase = state->phase;
         if (phase == 0)
         {
-            ((GameObject*)obj)->anim.modelState->flags &= ~DBSH_SYMBOL_OBJECT_MODEL_ACTIVE_FLAG;
+            ((GameObject*)obj)->anim.modelState->flags &= ~(u64)DBSH_SYMBOL_OBJECT_MODEL_ACTIVE_FLAG;
             state->phase = 1;
         }
         else if (phase == 2)
@@ -267,7 +267,7 @@ void dbsh_symbol_update(int obj)
         }
         else if (phase == 3)
         {
-            ((GameObject*)obj)->anim.modelState->flags &= ~DBSH_SYMBOL_OBJECT_MODEL_ACTIVE_FLAG;
+            ((GameObject*)obj)->anim.modelState->flags &= ~(u64)DBSH_SYMBOL_OBJECT_MODEL_ACTIVE_FLAG;
             if (state->flags.finished != 0)
             {
                 GameBit_Set(0x16b, 1);
